add point align test case
